document deployer key in hello world example
Commit
https://github.com/axelarnetwork/axelar-local-gmp-examples/commit/5d96e62eb30efd9fa614ae5562361e7fe255dfe0
updated the examples to read a deployer key from a config file. The README.md
in the https://github.com/axelarnetwork/axelar-local-gmp-examples repository
was updated, but the documentation of the "Hello World" example was not. This
commit updates the documentation of the "Hello World" example to more closely
match README.md in the example repository.