Update manual for bootstrapping onto testnet
In following these instructions, I found that the app would segfault with a snapshot downloaded from the given source on the old version (0.26.5), but with 0.28.0  I can start syncing a node with the network. 

I assume the same applies onto mainnet. I'm not sure what the status of casablanca is - LMK if I should update that as well.